+++ CHANGELOG.md
@@ -1,6 +1,5 @@
-# 4.2.0 (Nov 11, 2020)
-- Add notify model changed method (#1063)
-- Update to Kotlin 1.4.20-RC and remove dependency on kotlin-android-extensions
+#
+- Add support for `EpoxyModelGroup` in the `EpoxyVisibilityTracker` (#1091)
 
 # 4.1.0 (Sept 17, 2020)
 - Fix some synchronization issues with the parallel Epoxy processing option